Menambahkan validasi jika branch_id sudah ada di purchaseOrder branch tidak dapat di hapus

+++ app/Http/Controllers/BranchController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Branch;
+use App\Models\PurchaseOrder;
 use Illuminate\Http\Request;
 
 class BranchController extends Controller
@@ -48,6 +49,11 @@ class BranchController extends Controller
     {
         try {
             $branch = Branch::findBranch($id);
+
+            if (PurchaseOrder::where('branch_id', $id)->exists()) {
+                throw new \Exception('Cabang tidak bisa dihapus bila id branch sudah muncul di purchase_order!');
+            }
+
             Branch::deleteBranch($id);
             return redirect()->route('branch.list')->with('success', 'Cabang berhasil dihapus!');
         } catch (\Exception $e) {
